requirements.txt
gitdb==4.0.11
GitPython==3.1.41
pillow==10.3.0
setuptools==69.0.3
smmap==5.0.1
wheel==0.43.0
